+++ src/test/java/org/mmarini/wheelly/apis/RadarMapTest.java
@@ -27,14 +27,23 @@ package org.mmarini.wheelly.apis;
 
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.awt.geom.Point2D;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 import java.util.function.Predicate;
+import java.util.regex.Pattern;
+import java.util.stream.Stream;
 
+import static java.lang.String.format;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mmarini.Matchers.pointCloseTo;
+import static org.mmarini.Matchers.*;
 
 class RadarMapTest {
 
@@ -44,6 +53,101 @@ class RadarMapTest {
     public static final double MM1 = 0.001;
     public static final int MAX_INTERVAL = 10000;
     public static final int RECEPTIVE_ANGLE = 15;
+    public static final int ECHO_TIME = 100;
+
+    public static Stream<Arguments> findTargetDataset() {
+        // List<Point2D> obstacles, Point2D location, double maxDistance, Optional<Point2D> expected        return Stream.of(
+        return Stream.of(
+                parseRadarMap("""
+                        ...........
+                        ...........
+                        ...........
+                        ...........
+                        ...........
+                        .....X.....
+                        ...........
+                        ...........
+                        ...........
+                        ...........
+                        .....T....."""),
+                parseRadarMap("""
+                        ...........
+                        ...........
+                        ...........
+                        ...........
+                        ...........
+                        .....X.....
+                        ...........
+                        ...........
+                        ...........
+                        ..T........
+                        .....O....."""),
+                parseRadarMap("""
+                        .....T.....
+                        ...........
+                        ...........
+                        ...........
+                        ...O...O...
+                        ...O.X.O...
+                        ...O.O.O...
+                        ...........
+                        ...........
+                        ...........
+                        ..........."""),
+                parseRadarMap("""
+                        ...........
+                        ...........
+                        ...........
+                        ...........
+                        ....OOOOOOO
+                        .....X....O
+                        ..........O
+                        ..........O
+                        .T........O
+                        ...OOOOOOOO
+                        ...........""")
+        );
+    }
+
+    static Arguments parseRadarMap(String text) {
+        Point2D location = parseRadarMap(WIDTH, HEIGHT, GRID_SIZE, "X", text).getFirst();
+        Point2D target = parseRadarMap(WIDTH, HEIGHT, GRID_SIZE, "T", text).getFirst();
+        return Arguments.of(parseRadarMap(WIDTH, HEIGHT, GRID_SIZE, text), location, 1, target);
+    }
+
+    static List<Point2D> parseRadarMap(int width, int height, double size, String pattern, String text) {
+        List<Point2D> result = new ArrayList<>();
+        String[] lines = text.split("\n");
+        if (lines.length < height) {
+            throw new IllegalArgumentException(format("text must have %d line (%d", height, lines.length));
+        }
+        Predicate<String> p = Pattern.compile(pattern).asMatchPredicate();
+        for (int i = 0; i < width; i++) {
+            if (lines[i].length() < width) {
+                throw new IllegalArgumentException(format("lines %d must have size %d (%d)", i + 1, width, lines[i].length()));
+            }
+            double y = (height / 2 - i) * size;
+            for (int j = 0; j < width; j++) {
+                double x = (j - width / 2) * size;
+                String ch = lines[i].substring(j, j + 1);
+                if (p.test(ch)) {
+                    result.add(new Point2D.Double(x, y));
+                }
+            }
+        }
+        return result;
+    }
+
+    static RadarMap parseRadarMap(int width, int height, double size, String text) {
+        List<Point2D> echos = parseRadarMap(width, height, size, "O", text);
+        RadarMap radarMap = RadarMap.create(width, height, new Point2D.Double(), size,
+                MAX_INTERVAL, MAX_INTERVAL, MAX_INTERVAL, GRID_SIZE, RECEPTIVE_ANGLE);
+        for (Point2D p : echos) {
+            radarMap = radarMap.updateCell(radarMap.indexOf(p.getX(), p.getY()),
+                    c -> c.addEchogenic(ECHO_TIME));
+        }
+        return radarMap;
+    }
 
     @Test
     void cleanNoTimeout() {
@@ -76,7 +180,6 @@ class RadarMapTest {
 
     @Test
     void create() {
-
         RadarMap map = createRadarMap();
 
         assertEquals(GRID_SIZE, map.topology().gridSize());
@@ -98,6 +201,51 @@ class RadarMapTest {
     private RadarMap createRadarMap() {
         return RadarMap.create(WIDTH, HEIGHT, new Point2D.Double(), GRID_SIZE,
                 MAX_INTERVAL, MAX_INTERVAL, MAX_INTERVAL, GRID_SIZE, RECEPTIVE_ANGLE);
+    }
+
+    @ParameterizedTest
+    @MethodSource("findTargetDataset")
+    void findTargetTest(RadarMap map, Point2D location, double maxDistance, Point2D expected) {
+        // Given a radar map with obstacles
+        // And max distance
+        // And safe distance
+        double safeDistance = 0.15;
+
+        // When find target
+        Optional<Point2D> result = map.findTarget(location, maxDistance, safeDistance);
+
+        // Then ...
+        assertThat(result, expected != null
+                ? optionalOf(pointCloseTo(expected, 1e-3))
+                : emptyOptional());
+    }
+
+    @ParameterizedTest
+    @CsvSource({
+            // xFrom, yFrom, xTo, yTo, xObstacle, xObstacle, freeTrajectory
+            "0,0, 1,1, -0.4,-0.4, true",
+            "0,0, 1,1, 0.6,0.6, false",
+            "-0.4,-0.4, 0.688,0.688, 1,1, false",
+            "-0.4,-0.4, 0.687,0.687, 1,1, true",
+    })
+    void freeTrajectoryTest(double xFrom, double yFrom, double xTo, double yTo, double xObstacle,
+                            double yObstacle, boolean freeTrajectory) {
+        // Given a map with obstacle at xObstacle, yObstacle
+        RadarMap map = createRadarMap();
+        map = map.updateCell(
+                map.indexOf(xObstacle, yObstacle),
+                cell -> cell.addEchogenic(100)
+        );
+        // And the departure point
+        Point2D from = new Point2D.Double(xFrom, yFrom);
+        // And the destination point
+        Point2D to = new Point2D.Double(xTo, yTo);
+        // And safeDistance
+        double safeDistance = 0.3;
+
+        // When compute free trajectory
+        // Than ...
+        assertEquals(freeTrajectory, map.freeTrajectory(from, to, safeDistance));
     }
 
     @Test